Adding new test case for default password change after initial boot according to California law.
