+++ tests/platform_tests/test_first_time_boot_password_change/test_first_time_boot_password_change.py
@@ -5,13 +5,12 @@ Due to new law passed in California, each default user must change their default
 Important Note:
     Please run this test from sonic-mgmt/tests folder, otherwise it will fail.
 '''
-import pytest
-import logging
 import pexpect
 import time
-from tests.common.helpers.assertions import pytest_assert
+import pytest
 from tests.platform_tests.test_first_time_boot_password_change.default_consts import DefaultConsts
-from tests.platform_tests.test_first_time_boot_password_change.manufacture import manufacture
+from tests.platform_tests.test_first_time_boot_password_change.conftest import logger, currentConfigurations
+
 
 pytestmark = [
     pytest.mark.topology('any'),
@@ -19,88 +18,6 @@ pytestmark = [
     pytest.mark.disable_loganalyzer,
     pytest.mark.skip_check_dut_health
 ]
-
-
-class currentConfigurations:
-    '''
-    @summary: this class will act as a global database to save current configurations and changes the test made.
-    It will help us track the current state of the system,
-    and we will be used as part of cleanup fixtures.
-    '''
-    def __init__(self):
-        self.currentPassword = DefaultConsts.DEFAULT_PASSWORD  # initial password
-
-
-logger = logging.getLogger(__name__)
-currentConfigurations = currentConfigurations()
-
-
-@pytest.fixture(scope='module', autouse=True)
-def dut_hostname(request):
-    '''
-    @summary: this function returns the hostname of the dut from the 'host-pattern'
-    '''
-    hostname = request.config.getoption('--host-pattern')
-    logger.info("Hostname is {}".format(hostname))
-    return hostname
-
-
-@pytest.fixture(scope='module', autouse=True)
-def prepare_system_for_first_boot(request, dut_hostname):
-    '''
-    @summary: will manufacture the dut device to the given image in the parameter --restore_to_image,
-    by installing the image given from ONIE. for detailed information read the documentation
-    of the manufacture script.
-    '''
-    restore_image_path = request.config.getoption('restore_to_image')
-    pytest_assert(restore_image_path is not None, "restore_to_image param is empty, Please specify path to an image")
-    manufacture(dut_hostname, restore_image_path)
-
-
-def change_password(dut_hostname, username, current_password, new_password):
-    '''
-    @summary: this function changes the password for the user given
-    :param dut_hostname: host name of the dut
-    :param dut_ip: device under test
-    :param username: user name to change the password for
-    :param current_password: current password
-    :param new_password: new password
-    '''
-    logger.info("Changing password for username:{} to password: {}".format(username, new_password))
-    try:
-        # create a new ssh connection
-        engine = pexpect.spawn(DefaultConsts.SSH_COMMAND.format(username) + dut_hostname, timeout=15)
-        # because of race condition
-        engine.delaybeforesend = 0.2
-        engine.delayafterclose = 0.5
-        engine.expect(DefaultConsts.PASSWORD_REGEX)
-        engine.sendline(current_password + '\r')
-        engine.expect(DefaultConsts.SONIC_PROMPT)
-        engine.sendline('sudo usermod -p $(openssl passwd -1 {}) {}'.format(new_password, username) + '\r')
-        engine.expect(DefaultConsts.SONIC_PROMPT)
-        logger.info("Sleeping for {} secs to apply password change".format(DefaultConsts.APPLY_CONFIGURATIONS))
-        time.sleep(DefaultConsts.APPLY_CONFIGURATIONS)
-        engine.sendline('exit')
-        engine.close()
-    except Exception as err:
-        logger.info('Got an exception while changing the password')
-        logger.info(str(err))
-
-
-@pytest.fixture(scope='function', autouse=True)
-def restore_original_password(dut_hostname):
-    '''
-    @summary: this function will restore the original password to the default one to allow
-    the next test to use default password to login to dut.
-    '''
-    yield
-    logger.info("Sleep {} secs for system stabilization".format(DefaultConsts.STABILIZATION_TIME))
-    time.sleep(DefaultConsts.STABILIZATION_TIME)
-    logger.info("Restore original password")
-    change_password(dut_hostname,
-                    DefaultConsts.DEFAULT_USER,
-                    currentConfigurations.currentPassword,
-                    DefaultConsts.DEFAULT_PASSWORD)
 
 
 def test_default_password_change_after_first_boot(dut_hostname):
